programs/scoreboard_site/requirements.txt
streamlit
opencv-python-headless
ultralytics
pandas
Pillow
